+++ packages/cloud_firestore/cloud_firestore/lib/src/query.dart
@@ -186,6 +186,8 @@ abstract class Query<T extends Object?> {
     required FromFirestore<R> fromFirestore,
     required ToFirestore<R> toFirestore,
   });
+
+  AggregateQuery count();
 }
 
 /// Represents a [Query] over the data at a particular location.
@@ -196,7 +198,7 @@ class _JsonQuery implements Query<Map<String, dynamic>> {
     this.firestore,
     this._delegate,
   ) {
-    QueryPlatform.verifyExtends(_delegate);
+    QueryPlatform.verify(_delegate);
   }
 
   @override
@@ -808,6 +810,13 @@ class _JsonQuery implements Query<Map<String, dynamic>> {
 
   @override
   int get hashCode => Object.hash(runtimeType, firestore, _delegate);
+
+  /// Represents an [AggregateQuery] over the data at a particular location for retrieving metadata
+  /// without retrieving the actual documents.
+  @override
+  AggregateQuery count() {
+    return AggregateQuery._(_delegate.count(), this);
+  }
 }
 
 class _WithConverterQuery<T extends Object?> implements Query<T> {
@@ -970,4 +979,11 @@ class _WithConverterQuery<T extends Object?> implements Query<T> {
   @override
   int get hashCode =>
       Object.hash(runtimeType, _fromFirestore, _toFirestore, _originalQuery);
+
+  /// Represents an [AggregateQuery] over the data at a particular location for retrieving metadata
+  /// without retrieving the actual documents.
+  @override
+  AggregateQuery count() {
+    return _originalQuery.count();
+  }
 }