+++ packages/cloud_firestore_odm/cloud_firestore_odm/lib/annotation.dart
@@ -1,5 +1,47 @@
+// Copyright 2022, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 export 'src/validator.dart' show Min, Validator, Max;
 
+/// {@macro cloud_firestore_odm.named_query}
+class NamedQuery<T> {
+  /// {@template cloud_firestore_odm.named_query}
+  /// Defines a named query, allowing the ODM to generate utilities to interact
+  /// with the query in a type-safe way.
+  ///
+  /// By doing:
+  ///
+  /// ```dart
+  /// @NamedQuery<Person>('my-query-name')
+  /// @Collection<Person>(...)
+  /// class Anything {}
+  /// ```
+  ///
+  /// The ODM will generate a `myQueryNameGet` utility, which can be used as followed:
+  ///
+  /// ```dart
+  /// void main() async {
+  ///   Future<PersonSnapshot> snapshot = myQueryNameGet();
+  /// }
+  /// ```
+  ///
+  ///
+  /// **Note**:
+  /// Named queries **must** be associated with a [Collection] that has a
+  /// matching generic argument.
+  ///
+  /// This is necessary to ensure that `FirestoreDocumentSnapshot.reference` is
+  /// properly set.
+  ///
+  /// {@endtemplate}
+  const NamedQuery(this.queryName);
+
+  /// The name of the Firestore query that will be performed.
+  final String queryName;
+}
+
+/// {@template cloud_firestore_odm.collection}
 /// Defines a collection reference.
 ///
 /// To define a collection reference, first it is necessary to define a class
@@ -98,9 +140,14 @@ export 'src/validator.dart' show Min, Validator, Max;
 ///   FriendQuerySnapshot johnFriends = await johnRef.friends.get();
 /// }
 /// ```
+/// {@endtemplate}
 class Collection<T> {
+  /// {@macro cloud_firestore_odm.collection}
   const Collection(this.path, {this.name, this.prefix});
 
+  /// Decode a [Collection] from a [Map]
+  ///
+  /// This is internally used by the code-generator to decode configs from the `build.yaml`
   Collection.fromJson(Map<Object?, Object?> json)
       : this(
           json['path']! as String,
@@ -117,4 +164,43 @@ class Collection<T> {
 
   /// The prefix to use for generated class names. Defaults to the type of [T].
   final String? prefix;
+}
+
+/// {@macro cloud_firestore_odm.id}
+class Id {
+  /// {@template cloud_firestore_odm.id}
+  /// Marks a property as the document ID of a document.
+  ///
+  /// By default, the document ID is not present in the firestore object once decoded.
+  ///
+  /// While you can acccess it using the `DocumentSnapshot`, it isn't always convenient.
+  /// A solution to that is to use the `@Id` annotation, to tell Firestore that a
+  /// a given property in a class would be the document ID:
+  ///
+  /// ```dart
+  /// @Collection<Person>('users')
+  /// @firestoreSerializable
+  /// class Person {
+  ///   Person({
+  ///     required this.name,
+  ///     required this.age,
+  ///     required this.id,
+  ///   });
+  ///
+  ///   // By adding this annotation, this property will not be considered as part
+  ///   // of the Firestore document, but instead represent the document ID.
+  ///   @Id()
+  ///   final String id;
+  ///
+  ///   final String name;
+  ///   final int age;
+  /// }
+  /// ```
+  ///
+  /// There are a few restrictions when using this annotation:
+  ///
+  /// - It can be used only once within an object
+  /// - The annotated property must be of type `String`.
+  /// {@endtemplate}
+  const Id();
 }