+++ packages/cloud_functions/cloud_functions_platform_interface/lib/src/platform_interface/platform_interface_https_callable.dart
@@ -6,6 +6,7 @@
 import 'dart:async';
 
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+
 import '../../cloud_functions_platform_interface.dart';
 
 /// Interface for [HttpsCallable] implementations.
@@ -24,8 +25,8 @@ abstract class HttpsCallablePlatform extends PlatformInterface {
   /// This is used by the app-facing [HttpsCallable] to ensure that
   /// the object in which it's going to delegate calls has been
   /// constructed properly.
-  static void verifyExtends(HttpsCallablePlatform instance) {
-    PlatformInterface.verifyToken(instance, _token);
+  static void verify(HttpsCallablePlatform instance) {
+    PlatformInterface.verify(instance, _token);
   }
 
   /// The [FirebaseFunctionsPlatform] instance.