+++ packages/firebase_app_check/firebase_app_check/example/lib/main.dart
@@ -1,7 +1,12 @@
+// Copyright 2022, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 import 'package:firebase_app_check/firebase_app_check.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 
 import 'firebase_options.dart';
 
@@ -17,6 +22,7 @@ Future<void> main() async {
       // Your personal reCaptcha public key goes here:
       .activate(
     webRecaptchaSiteKey: kWebRecaptchaSiteKey,
+    androidDebugProvider: true,
   );
 
   runApp(MyApp());
@@ -81,6 +87,25 @@ class _FirebaseAppCheck extends State<FirebaseAppCheckExample> {
           children: [
             ElevatedButton(
               onPressed: () async {
+                // Use this button to check whether the request was validated on the Firebase console
+                // Gets first document in collection
+                final result = await FirebaseFirestore.instance
+                    .collection('flutter-tests')
+                    .limit(1)
+                    .get();
+
+                if (result.docs.isNotEmpty) {
+                  setMessage('Document found');
+                } else {
+                  setMessage(
+                    'Document not found, please add a document to the collection',
+                  );
+                }
+              },
+              child: const Text('Test App Check validates requests'),
+            ),
+            ElevatedButton(
+              onPressed: () async {
                 if (kIsWeb) {
                   print(
                     'Pass in your "webRecaptchaSiteKey" key found on you Firebase Console to activate if using on the web platform.',
@@ -88,6 +113,7 @@ class _FirebaseAppCheck extends State<FirebaseAppCheckExample> {
                 }
                 await appCheck.activate(
                   webRecaptchaSiteKey: kWebRecaptchaSiteKey,
+                  androidDebugProvider: true,
                 );
                 setMessage('activated!!');
               },