+++ packages/firebase_auth/firebase_auth_platform_interface/test/auth_credential_test.dart
@@ -53,7 +53,7 @@ void main() {
 
       expect(
         result,
-        'AuthCredential(providerId: $kMockProviderId, signInMethod: $kMockSignInMethod, token: $kMockToken)',
+        'AuthCredential(providerId: $kMockProviderId, signInMethod: $kMockSignInMethod, token: $kMockToken, accessToken: null)',
       );
     });
   });