+++ packages/firebase_auth/firebase_auth_web/lib/src/utils/web_utils.dart
@@ -80,6 +80,7 @@ FirebaseAuthException getFirebaseAuthException(
         MultiFactorSessionWeb('web', resolverWeb.session),
         FirebaseAuthWeb.instance,
         resolverWeb,
+        auth,
       ),
     );
   }
@@ -278,16 +279,25 @@ AuthCredential? convertWebAuthCredential(
 
 /// Converts a [auth_interop.OAuthCredential] into a [AuthCredential].
 AuthCredential? convertWebOAuthCredential(
-  auth_interop.OAuthCredential? oAuthCredential,
+  auth_interop.UserCredential? userCredential,
 ) {
-  if (oAuthCredential == null) {
+  if (userCredential == null) {
     return null;
   }
 
-  return OAuthProvider(oAuthCredential.providerId).credential(
-    accessToken: oAuthCredential.accessToken,
-    secret: oAuthCredential.secret,
-    idToken: oAuthCredential.idToken,
+  final authCredential = auth_interop.OAuthProvider.credentialFromResult(
+    userCredential.jsObject,
+  );
+
+  if (authCredential == null) {
+    return null;
+  }
+
+  return OAuthProvider(authCredential.providerId).credential(
+    signInMethod: authCredential.signInMethod,
+    accessToken: authCredential.accessToken,
+    secret: authCredential.secret,
+    idToken: authCredential.idToken,
   );
 }