+++ packages/firebase_database/firebase_database/lib/src/database_event.dart
@@ -10,7 +10,7 @@ class DatabaseEvent {
   final DatabaseEventPlatform _delegate;
 
   DatabaseEvent._(this._delegate) {
-    DatabaseEventPlatform.verifyExtends(_delegate);
+    DatabaseEventPlatform.verify(_delegate);
   }
 
   /// The type of event.