+++ packages/firebase_dynamic_links/firebase_dynamic_links_platform_interface/lib/src/method_channel/utils/exception.dart
@@ -5,45 +5,15 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/services.dart';
 
+import 'package:_flutterfire_internals/_flutterfire_internals.dart';
+
 /// Catches a [PlatformException] and returns an [Exception].
 ///
 /// If the [Exception] is a [PlatformException], a [FirebaseException] is returned.
 Never convertPlatformException(Object exception, StackTrace stackTrace) {
-  if (exception is! Exception || exception is! PlatformException) {
-    Error.throwWithStackTrace(exception, stackTrace);
-  }
-
-  Error.throwWithStackTrace(
-    platformExceptionToFirebaseException(exception, stackTrace),
+  convertPlatformExceptionToFirebaseException(
+    exception,
     stackTrace,
-  );
-}
-
-/// Converts a [PlatformException] into a [FirebaseException].
-///
-/// A [PlatformException] can only be converted to a [FirebaseException] if the
-/// `details` of the exception exist. Firebase returns specific codes and messages
-/// which can be converted into user friendly exceptions.
-FirebaseException platformExceptionToFirebaseException(
-  PlatformException platformException,
-  StackTrace? stackTrace,
-) {
-  Map<String, String>? details = platformException.details != null
-      ? Map<String, String>.from(platformException.details)
-      : null;
-
-  String code = 'unknown';
-  String message = platformException.message ?? '';
-
-  if (details != null) {
-    code = details['code'] ?? code;
-    message = details['message'] ?? message;
-  }
-
-  return FirebaseException(
     plugin: 'firebase_dynamic_links',
-    code: code,
-    message: message,
-    stackTrace: stackTrace,
   );
 }