+++ packages/firebase_messaging/firebase_messaging_platform_interface/lib/src/platform_interface/platform_interface_messaging.dart
@@ -56,7 +56,7 @@ abstract class FirebaseMessagingPlatform extends PlatformInterface {
 
   /// Sets the [FirebaseMessagingPlatform.instance]
   static set instance(FirebaseMessagingPlatform instance) {
-    PlatformInterface.verifyToken(instance, _token);
+    PlatformInterface.verify(instance, _token);
     _instance = instance;
   }
 
@@ -304,5 +304,16 @@ abstract class FirebaseMessagingPlatform extends PlatformInterface {
   /// Unsubscribe from topic in background.
   Future<void> unsubscribeFromTopic(String topic) {
     throw UnimplementedError('unsubscribeFromTopic() is not implemented');
+  }
+
+  /// Enables or disables Firebase Cloud Messaging message delivery metrics export to BigQuery.
+  ///
+  /// On iOS, you need to follow [this guide](https://firebase.google.com/docs/cloud-messaging/understand-delivery?platform=ios#enable_delivery_data_export_for_background_notifications)
+  /// in order to export metrics to BigQuery.
+  /// On Web, you need to setup a [service worker](https://firebase.google.com/docs/cloud-messaging/js/client) and call `experimentalSetDeliveryMetricsExportedToBigQueryEnabled(messaging, true)`
+  Future<void> setDeliveryMetricsExportToBigQuery(bool enabled) {
+    throw UnimplementedError(
+      'setDeliveryMetricsExportToBigQuery() is not implemented',
+    );
   }
 }