+++ packages/firebase_messaging/firebase_messaging_web/lib/src/interop/messaging_interop.dart
@@ -8,8 +8,8 @@
 @JS('firebase_messaging')
 library firebase_interop.messaging;
 
-import 'package:js/js.dart';
 import 'package:firebase_core_web/firebase_core_web_interop.dart';
+import 'package:js/js.dart';
 
 @JS()
 external MessagingJsImpl getMessaging([AppJsImpl? app]);