+++ packages/cloud_functions/cloud_functions_platform_interface/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 5.1.21
+
+ - Update a dependency to the latest release.
+
+## 5.1.20
+
+ - Update a dependency to the latest release.
+
+## 5.1.19
+
+ - Update a dependency to the latest release.
+
+## 5.1.18
+
+ - Update a dependency to the latest release.
+
 ## 5.1.17
 
  - Update a dependency to the latest release.