+++ packages/firebase_analytics/firebase_analytics/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 10.0.2
+
+ - Update a dependency to the latest release.
+
+## 10.0.1
+
+- Update a dependency to the latest release.
+
+## 10.0.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FEAT**: Firebase iOS SDK version: `10.0.0` ([#9708](https://github.com/firebase/flutterfire/issues/9708)). ([9627c56a](https://github.com/firebase/flutterfire/commit/9627c56a37d657d0250b6f6b87d0fec1c31d4ba3))
+
+## 9.3.8
+
+ - Update a dependency to the latest release.
+
+## 9.3.7
+
+ - Update a dependency to the latest release.
+
 ## 9.3.6
 
  - **FIX**: suppress unchecked warnings that aren't necessary ([#9532](https://github.com/firebase/flutterfire/issues/9532)). ([3ebd4593](https://github.com/firebase/flutterfire/commit/3ebd4593d11fbbd359b8d514a9c0577654859992))