+++ packages/firebase_app_installations/firebase_app_installations/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.2.1+1
+
+ - Update a dependency to the latest release.
+
+## 0.2.1
+
+- Update a dependency to the latest release.
+
+## 0.2.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FEAT**: Firebase iOS SDK version: `10.0.0` ([#9708](https://github.com/firebase/flutterfire/issues/9708)). ([9627c56a](https://github.com/firebase/flutterfire/commit/9627c56a37d657d0250b6f6b87d0fec1c31d4ba3))
+
+## 0.1.1+10
+
+ - Update a dependency to the latest release.
+
+## 0.1.1+9
+
+ - Update a dependency to the latest release.
+
 ## 0.1.1+8
 
  - Update a dependency to the latest release.