+++ packages/firebase_dynamic_links/firebase_dynamic_links_platform_interface/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.2.3+18
+
+ - Update a dependency to the latest release.
+
+## 0.2.3+17
+
+ - Update a dependency to the latest release.
+
+## 0.2.3+16
+
+ - Update a dependency to the latest release.
+
+## 0.2.3+15
+
+ - Update a dependency to the latest release.
+
+## 0.2.3+14
+
+ - Update a dependency to the latest release.
+
 ## 0.2.3+13
 
  - Update a dependency to the latest release.