+++ packages/firebase_messaging/firebase_messaging_web/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 3.2.3
+
+ - Update a dependency to the latest release.
+
+## 3.2.2
+
+ - Update a dependency to the latest release.
+
+## 3.2.1
+
+ - Update a dependency to the latest release.
+
+## 3.2.0
+
+ - **FEAT**: add support for exporting delivery metrics to BigQuery ([#9636](https://github.com/firebase/flutterfire/issues/9636)). ([170b99b9](https://github.com/firebase/flutterfire/commit/170b99b91573f28316172e43188d57ca14600446))
+
+## 3.1.6
+
+ - Update a dependency to the latest release.
+
 ## 3.1.5
 
  - Update a dependency to the latest release.